separate parsed rfml test logic
